Rename lhashlib to lcryptolib, preload the library so we don't need to worry about name collisions.
