+++ src/linit.cpp
@@ -50,7 +50,12 @@ static const luaL_Reg loadedlibs[] = {
   {LUA_MATHLIBNAME, luaopen_math},
   {LUA_UTF8LIBNAME, luaopen_utf8},
   {LUA_DBLIBNAME, luaopen_debug},
-  {LUA_HASHLIBNAME, luaopen_hashlib},
+  {NULL, NULL}
+};
+
+
+static const luaL_Reg preloadedLibs[] = {
+  {"crypto", luaopen_crypto},
   {NULL, NULL}
 };
 
@@ -58,11 +63,19 @@ static const luaL_Reg loadedlibs[] = {
 LUALIB_API void luaL_openlibs (lua_State *L)
 {
   const luaL_Reg *lib;
-  /* "require" functions from 'loadedlibs' and set results to global table */
+
   for (lib = loadedlibs; lib->func; lib++)
   {
     luaL_requiref(L, lib->name, lib->func, 1);
     lua_pop(L, 1);  /* remove lib */
+  }
+
+  for (lib = preloadedLibs; lib->func; lib++)
+  {  
+    luaL_getsubtable(L, LUA_REGISTRYINDEX, LUA_PRELOAD_TABLE);
+    lua_pushcfunction(L, lib->func);
+    lua_setfield(L, -2, lib->name);
+    lua_pop(L, 1);
   }
 }
