+++ src/lualib.h
@@ -41,8 +41,8 @@ LUAMOD_API int (luaopen_debug) (lua_State *L);
 #define LUA_LOADLIBNAME	"package"
 LUAMOD_API int (luaopen_package) (lua_State *L);
 
-#define LUA_HASHLIBNAME "hashlib"
-LUAMOD_API int (luaopen_hashlib) (lua_State *L);
+
+LUAMOD_API int (luaopen_crypto) (lua_State *L);
 
 
 /* open all previous libraries */
