+++ tests/basic.lua
@@ -549,9 +549,10 @@ end
 
 print "Testing standard library additions."
 do
-    assert(hashlib.fnv1("hello world") == 0x7DCF62CDB1910E6F)
-    assert(hashlib.fnv1a("hello world") == 8618312879776256743)
-    assert(hashlib.joaat("hello world") == 1045060183)
+    local crypto = require("crypto")
+    assert(crypto.fnv1("hello world") == 0x7DCF62CDB1910E6F)
+    assert(crypto.fnv1a("hello world") == 8618312879776256743)
+    assert(crypto.joaat("hello world") == 1045060183)
 end
 do
     local t = {}
